Removed some wildcart imports and removed the color setting
