--- pokete_classes/__init__.py
@@ -1,7 +1,0 @@
-from pokete_classes.classes import *
-from pokete_classes.ui_elements import *
-from pokete_classes.effects import effects
-
-
-if __name__ == "__main__":
-    print("\033[31;1mDo not execute this!\033[0m")
